权限

+++ onebase-module-app/onebase-module-app-api/src/main/java/com/cmsr/onebase/module/app/api/app/AppApplicationApi.java
@@ -19,7 +19,7 @@ public interface AppApplicationApi {
 
     List<ApplicationDTO> findAppApplicationByAppIds(Collection<Long> appIds);
 
-    Map<Integer,Integer> findAppApplicationAll();
+    Map<Long,Integer> findAppApplicationAll();
 
     void  updateAppTimeById(Long appId);
